added global window hooks installation
This is the first step in moving towards improved gui agent architecture.
Currently wga receives damage notifications from the qvideo driver, but
needs to basically poll all active windows to check what changed. This
is obviously very bad for performance, but also can miss some subtle
changes in window states. Shell hook that is currently used for receiving
notifications only receives a small fraction of what the agent actually
needs to know (eg. no "window was minimized" notifications).

Installing global message hooks will give us the ability to track every
window message in the system. It will eliminate the need for polling
window states because we will track window state changes as they happen.
We will also know exactly which windows were repainted.

We need two hook DLLs to be able to monitor all processes on a 64-bit system,
one for each architecture (or bitness). A 64-bit DLL can only be injected
into 64-bit processes, and the same goes for 32 bits. This also implies that
we also need a 32-bit host process to initiate the 32-bit hooks, as the main
64-bit wga process can't load 32-bit hook DLL. wga executes the 32-bit host
process at the start of its main event loop and terminates it when it exits.

Hook DLLs are basically just transparent message "proxies" -- when the
appropriate hook procedure receives a message it's interested in, it packs
the message into a "bitness-neutral" packet and sends that to wga through
an IPC channel (currently implemented as a mailslot). wga listens to the IPC
messages in the main event loop by an asynchronous read and processes
messages as they arrive.